[WebDriver BiDi] Fix fetch cookie network test.
