+++ webdriver/tests/bidi/integration/cookies_with_network_events.py
@@ -1,7 +1,10 @@
 import pytest
 
 from webdriver.bidi.modules.script import ContextTarget
-from webdriver.bidi.modules.storage import BrowsingContextPartitionDescriptor
+from webdriver.bidi.modules.storage import (
+    BrowsingContextPartitionDescriptor,
+    StorageKeyPartitionDescriptor,
+)
 
 from .. import assert_cookies
 
@@ -110,6 +113,7 @@ async def test_fetch(
     fetch,
     wait_for_future_safe,
     url,
+    origin,
     domain_1,
 ):
     # Clean up cookies in case some other tests failed before cleaning up.
@@ -124,9 +128,10 @@ async def test_fetch(
 
     cookie_name = "foo"
     cookie_value = "bar"
+    path = "/webdriver/tests/support/http_handlers"
     # Add `Access-Control-Allow-Origin` header for cross-origin request to work.
     request_url = url(
-        "/webdriver/tests/support/http_handlers/headers.py?header=Access-Control-Allow-Origin:*",
+        f"{path}/headers.py?header=Access-Control-Allow-Origin:*",
         domain=domain_1,
     )
 
@@ -145,7 +150,8 @@ async def test_fetch(
     await wait_for_future_safe(on_before_request_sent)
 
     result = await bidi_session.storage.get_cookies(
-        partition=BrowsingContextPartitionDescriptor(new_tab["context"])
+        partition=StorageKeyPartitionDescriptor(source_origin=origin(domain=domain_1)),
+        filter={"path": path},
     )
     assert_cookies(result["cookies"], events[0]["request"]["cookies"])
